Adds datalistId for prompt modals
The datalistId points to an id of a <datalist> element. The options
f this datalist element are used as the list attribute in the
input> element of a prompt dialog.  Use it like this:

In your component template:

    <datalist id="my-list"><option value="1"><option value="2"></datalist>

In your component js:

   modal.prompt()
     .datalistId('my-list')
     .size('lg')
     .isBlocking(true)
      .showClose(true)
     .keyboard(27)
     .title('Hello World')
     .body('A Customized Modal')
     .open();

This adds options 1 and 2 as elements of the modal.